Order model - serializing products

+++ app/Http/Controllers/CheckoutController.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use DateTime;
 use Stripe\Stripe;
+use App\Models\Order;
 use Stripe\PaymentIntent;
 use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 use Gloudemans\Shoppingcart\Facades\Cart;
 
 class CheckoutController extends Controller
@@ -44,11 +47,47 @@ class CheckoutController extends Controller
 
     public function store(Request $request)
     {
-        Cart::destroy();
+        
 
         $data = $request->json()->all();
 
-        return $data['paymentIntent'];
+        $order = new Order();
+
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+        $order->amount = $data['paymentIntent']['amount'];
+        $order->payment_created_at = (new DateTime())
+                                        ->setTimestamp($data['paymentIntent']['created'])
+                                        ->format('Y-m-d H:i:s');
+
+
+        $products = [];
+        $i = 0;
+
+        foreach(Cart::content() as $product){
+            $products['product_' . $i][] = $product->model->title;
+            $products['product_' . $i][] = $product->model->price;
+            $products['product_' . $i][] = $product->qty;
+            $i++;
+        }
+
+        $order->products = \serialize($products);
+        $order->user_id = 15;
+        $order->save();
+
+        if($data['paymentIntent']['status'] == 'succeeded'){
+            Cart::destroy();
+            Session::flash('success','votre commande a été traitée avec succés');
+            return \response()->json(['sccess' => 'Payment Intent Succeeded']);
+        }else{
+            return \response()->json(['error' => 'Payment Intent Not Succeeded']);
+        }
+
+
+
+    }
+
+    public function thankYou(){
+        return Session::has('success') ? view('checkout.thanks') : redirect()->route('products.index');
     }
